# docs/protocol/openapi.yaml
openapi: 3.1.0
info:
  title: QuanCard Server API
  version: 1.0.0
  license:
    name: AGPL-3.0-only
    identifier: AGPL-3.0-only
  description: 'HTTP API implemented by the server. All JSON request objects described here reject unknown properties.
    Routes without a requestBody do not inspect a parsed body. Binary revision uploads contain an opaque sync-v1
    envelope JSON object, not a JSON API wrapper. Base64 fields use the standard padded alphabet. Timestamps are
    Unix seconds.


    API requests require HTTPS and a Host matching the configured public origin, with an explicitly configured localhost
    development exception. GET /healthz bypasses these checks. Non-Bearer API writes require Origin matching the
    allowed origin and X-QuanCard-Client: web, including unauthenticated setup and login; /api/v1/pairings/claim
    is exempt. The Bearer-prefix exemption is a transport guard and does not grant authentication or access.


    Session cookies are host-only, Secure, HttpOnly and SameSite=Strict. Device bearer tokens begin with qcd_ and
    can only access manifest and revisions of their paired vault. A recognized qcd_ bearer credential takes precedence
    over a session cookie. API responses use Cache-Control: no-store. Errors use Error and HTTP 429 also includes
    Retry-After. HEAD counterparts are automatically exposed and return no body.


    When webRoot is configured, the server additionally serves deployment-dependent static files and returns its
...
servers:
...
paths:
...
  /api/v1/vaults:
    get:
      operationId: listVaults
      summary: List the account's vaults
      description: Requires a session. A device token cannot list vaults or retrieve wrapped vault keys through
        this endpoint.
      security:
      - session: []
      responses:
        '200':
          description: Vaults ordered by creation time.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                - vaults
                properties:
                  vaults:
                    type: array
                    items:
                      $ref: '#/components/schemas/Vault'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '421':
          $ref: '#/components/responses/MisdirectedRequest'
        '426':
          $ref: '#/components/responses/HttpsRequired'
        '500':
          $ref: '#/components/responses/Internal'
    post:
      operationId: createVault
      summary: Create a vault
      description: 'JSON request limit: 64 KiB. Validates the outer manifest and wrapped-key envelopes. An identical
        retry returns 200, otherwise creation returns 201. The account vault quota is checked before idempotency,
        so a retry at capacity returns 507.'
      security:
      - session: []
      parameters:
      - $ref: '#/components/parameters/Origin'
      - $ref: '#/components/parameters/WebClient'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVaultRequest'
      responses:
        '200':
          description: Identical existing vault creation acknowledged.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vault'
        '201':
          description: Vault created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vault'
        '400':
          description: 'badRequest: malformed fields, UUID or Base64; invalidEnvelope: manifest/key envelope malformed,
            oversized or bound to the wrong kind or vaultID.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/CrossOriginRejected'
        '409':
          description: 'vaultExists: vaultID already exists with a different owner, manifest or wrapped key.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '421':
          $ref: '#/components/responses/MisdirectedRequest'
        '426':
          $ref: '#/components/responses/HttpsRequired'
        '500':
          $ref: '#/components/responses/Internal'
        '507':
          description: 'capacity: the account already has the configured maximum number of vaults.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    head:
      operationId: headListVaults
      summary: 'Get headers: List the account''s vaults'
      description: Automatically exposed by Fastify for this GET route. Uses the same handler and authentication,
        but suppresses response bodies, including Error bodies.
      security:
      - session: []
      responses:
        '200':
          description: Vaults ordered by creation time. HEAD responses have no body.
        '401':
          description: 'unauthorized: no valid session, or a device token used on a session-only endpoint. HEAD
            responses have no body.'
        '421':
          description: 'misdirectedRequest: Host does not match the configured public origin (except permitted local
            development requests). HEAD responses have no body.'
        '426':
          description: 'httpsRequired: HTTPS is required; only explicitly enabled localhost development is exempt.
            HEAD responses have no body.'
          headers:
            Connection:
              schema:
                type: string
                const: close
              description: The server closes the insecure connection.
        '500':
          description: 'internal: unexpected server failure. HEAD responses have no body.'
  /api/v1/vaults/{vaultID}:
    delete:
      operationId: deleteVault
      summary: Delete a vault
      description: Requires the owning account's session and fresh authentication. Removes the vault, revisions, pairings
        and paired devices.
      security:
      - session: []
      parameters:
      - $ref: '#/components/parameters/Origin'
      - $ref: '#/components/parameters/WebClient'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FreshAuthRequest'
      responses:
        '204':
          description: Vault and dependent records deleted; no response body.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: 'unauthorized: a valid session is required. invalidCredentials: the supplied authentication
            key is incorrect.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          $ref: '#/components/responses/CrossOriginRejected'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '421':
          $ref: '#/components/responses/MisdirectedRequest'
        '426':
          $ref: '#/components/responses/HttpsRequired'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
    parameters:
    - $ref: '#/components/parameters/vaultID'
  /api/v1/vaults/{vaultID}/manifest:
    get:
      operationId: getManifest
      summary: Download the vault manifest
      description: Accepts the owning account's session or a device token scoped to this vault. An out-of-scope
        vault returns notFound.
      security:
      - session: []
      - device: []
      responses:
        '200':
          description: Stored opaque sync-v1 manifest envelope JSON bytes; not Base64.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
                maxLength: 8192
                description: Raw envelope JSON bytes, objectKind syncManifest.v1, recordID equal to vaultID.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '421':
          $ref: '#/components/responses/MisdirectedRequest'
        '426':
          $ref: '#/components/responses/HttpsRequired'
        '500':
          $ref: '#/components/responses/Internal'
    parameters:
    - $ref: '#/components/parameters/vaultID'
    head:
      operationId: headGetManifest
      summary: 'Get headers: Download the vault manifest'
      description: Automatically exposed by Fastify for this GET route. Uses the same handler and authentication,
        but suppresses response bodies, including Error bodies.
      security:
      - session: []
      - device: []
      responses:
        '200':
          description: Stored opaque sync-v1 manifest envelope JSON bytes; not Base64. HEAD responses have no body.
        '400':
          description: 'badRequest: malformed request, invalid field type or value, unknown JSON body property,
            invalid UUID or request parsing error. HEAD responses have no body.'
        '401':
          description: 'unauthorized: no valid session, or a device token used on a session-only endpoint. HEAD
            responses have no body.'
        '404':
          description: 'notFound: resource does not exist, belongs to another account, or is outside the device
            token''s vault. HEAD responses have no body.'
        '421':
          description: 'misdirectedRequest: Host does not match the configured public origin (except permitted local
            development requests). HEAD responses have no body.'
        '426':
          description: 'httpsRequired: HTTPS is required; only explicitly enabled localhost development is exempt.
            HEAD responses have no body.'
          headers:
            Connection:
              schema:
                type: string
                const: close
              description: The server closes the insecure connection.
        '500':
          description: 'internal: unexpected server failure. HEAD responses have no body.'
  /api/v1/vaults/{vaultID}/revisions:
    get:
      operationId: listRevisions
      summary: Page through stored revisions
      description: Accepts the owning account's session or a device token scoped to this vault. Returns sequences
        strictly greater than after, in ascending order. A page holds at most 500 revisions and 24 MiB of decoded
        envelope bytes; the first record is always included if present. Unknown query parameters are ignored.
      security:
      - session: []
      - device: []
      parameters:
      - name: after
        in: query
        description: Exclusive sequence cursor. Converted with Number and must be a non-negative safe integer.
        schema:
          type: integer
          minimum: 0
          maximum: 9007199254740991
          default: 0
      - name: limit
        in: query
        description: Requested page size, default 500. Converted with Number and capped at 500 before positive-safe-integer
          validation; values above 500 are accepted and capped.
        schema:
          type: integer
          minimum: 1
          default: 500
      responses:
        '200':
          description: Ascending sequence page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevisionPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '421':
          $ref: '#/components/responses/MisdirectedRequest'
        '426':
          $ref: '#/components/responses/HttpsRequired'
        '500':
          $ref: '#/components/responses/Internal'
    parameters:
    - $ref: '#/components/parameters/vaultID'
    head:
      operationId: headListRevisions
      summary: 'Get headers: Page through stored revisions'
      description: Automatically exposed by Fastify for this GET route. Uses the same handler and authentication,
        but suppresses response bodies, including Error bodies.
      security:
      - session: []
      - device: []
      parameters:
      - name: after
        in: query
        description: Exclusive sequence cursor. Converted with Number and must be a non-negative safe integer.
        schema:
          type: integer
          minimum: 0
          maximum: 9007199254740991
          default: 0
      - name: limit
        in: query
        description: Requested page size, default 500. Converted with Number and capped at 500 before positive-safe-integer
          validation; values above 500 are accepted and capped.
        schema:
          type: integer
          minimum: 1
          default: 500
      responses:
        '200':
          description: Ascending sequence page. HEAD responses have no body.
        '400':
          description: 'badRequest: malformed request, invalid field type or value, unknown JSON body property,
            invalid UUID or request parsing error. HEAD responses have no body.'
        '401':
          description: 'unauthorized: no valid session, or a device token used on a session-only endpoint. HEAD
            responses have no body.'
        '404':
          description: 'notFound: resource does not exist, belongs to another account, or is outside the device
            token''s vault. HEAD responses have no body.'
        '421':
          description: 'misdirectedRequest: Host does not match the configured public origin (except permitted local
            development requests). HEAD responses have no body.'
        '426':
          description: 'httpsRequired: HTTPS is required; only explicitly enabled localhost development is exempt.
            HEAD responses have no body.'
          headers:
            Connection:
              schema:
                type: string
                const: close
              description: The server closes the insecure connection.
        '500':
          description: 'internal: unexpected server failure. HEAD responses have no body.'
  /api/v1/vaults/{vaultID}/revisions/{revisionID}:
    put:
      operationId: putRevision
      summary: Store an immutable revision
      description: 'Upload raw opaque sync-v1 envelope JSON bytes, at most 16 MiB (16777216 bytes). Content-Type
        must be application/octet-stream. The server validates only the outer envelope: objectKind syncRevision.v1
        and recordID equal to revisionID. Encrypted vault identity and revision contents cannot be inspected by
        the server. Identical bytes are idempotent and are acknowledged before quota checks. Device tokens may upload
        only to their paired vault; Bearer requests are exempt from browser-write headers.'
      security:
      - session: []
      - device: []
      parameters:
      - $ref: '#/components/parameters/Origin'
      - $ref: '#/components/parameters/WebClient'
      responses:
        '200':
          description: Identical retry acknowledged with the original sequence.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevisionReceipt'
        '201':
          description: Revision stored with a new sequence.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevisionReceipt'
        '400':
          description: 'badRequest: invalid path UUID or request parsing failure; invalidEnvelope: malformed envelope,
            wrong objectKind or recordID, or oversized envelope detected during envelope validation.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/CrossOriginRejected'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: 'revisionConflict: this revisionID already has different bytes in the vault.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '421':
          $ref: '#/components/responses/MisdirectedRequest'
        '426':
          $ref: '#/components/responses/HttpsRequired'
        '500':
          $ref: '#/components/responses/Internal'
        '507':
          description: 'capacity: adding this revision would exceed the configured per-vault revision count or total-byte
            quota.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        description: Opaque sync-v1 envelope JSON bytes; maximum 16 MiB. Send the bytes directly, without an outer
          JSON wrapper or Base64 encoding.
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
              maxLength: 16777216
              contentMediaType: application/json
              contentSchema:
                $ref: '#/components/schemas/RevisionEnvelope'
              description: Raw UTF-8 envelope JSON bytes. The decoded JSON object has exactly the properties in
                RevisionEnvelope.
    parameters:
    - $ref: '#/components/parameters/vaultID'
    - $ref: '#/components/parameters/revisionID'
  /api/v1/vaults/{vaultID}/pairings:
    post:
      operationId: createPairing
      summary: Create a device pairing code
      description: Requires the owning account's session and fresh authentication. The browser adds the vault key
        to its QR code; this endpoint never returns that key.
      security:
      - session: []
      parameters:
      - $ref: '#/components/parameters/Origin'
      - $ref: '#/components/parameters/WebClient'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FreshAuthRequest'
      responses:
        '201':
          description: Single-use 32-character URL-safe code created; expires after 10 minutes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodeWithExpiry'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: 'unauthorized: a valid session is required. invalidCredentials: the supplied authentication
            key is incorrect.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          $ref: '#/components/responses/CrossOriginRejected'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '421':
          $ref: '#/components/responses/MisdirectedRequest'
        '426':
          $ref: '#/components/responses/HttpsRequired'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
    parameters:
    - $ref: '#/components/parameters/vaultID'
  /api/v1/pairings/claim:
    post:
      operationId: claimPairing
      summary: Exchange a pairing code for a device token
      description: Native endpoint exempt from Origin and X-QuanCard-Client checks even without Bearer credentials.
        Limited to 10 attempts per minute per client. The returned token can only access manifest and revision endpoints
        of the paired vault.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimPairingRequest'
      responses:
        '201':
          description: Pairing consumed and device credential issued.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                - deviceID
                - vaultID
                - deviceToken
                properties:
                  deviceID:
                    $ref: '#/components/schemas/UUID'
                  vaultID:
                    $ref: '#/components/schemas/UUID'
                  deviceToken:
                    type: string
                    pattern: ^qcd_[A-Za-z0-9_-]{43}$
                    description: Bearer credential, returned only upon claim.
        '400':
          $ref: '#/components/responses/BadRequest'
        '410':
          description: 'pairingInvalid: code is missing, malformed, unknown, expired or already claimed.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '421':
          $ref: '#/components/responses/MisdirectedRequest'
        '426':
          $ref: '#/components/responses/HttpsRequired'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: __Host-qc_session
      description: Browser session cookie. Device tokens cannot substitute for a session on account, vault-management
        or device-management endpoints.
    device:
      type: http
      scheme: bearer
      bearerFormat: qcd_<opaque token>
      description: Opaque device token issued by pairing claim. Only accepts manifest GET/HEAD, revisions GET/HEAD
        and revision PUT for its single paired vault.
...
  responses:
...
    Unauthorized:
      description: 'unauthorized: no valid session, or a device token used on a session-only endpoint.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    MisdirectedRequest:
      description: 'misdirectedRequest: Host does not match the configured public origin (except permitted local
        development requests).'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Internal:
      description: 'internal: unexpected server failure.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...